dry preprocessor tests
