+++ packages/houdini-preprocess/src/utils/index.ts
@@ -1,4 +1,5 @@
 export { default as walkTaggedDocuments } from './walkTaggedDocuments'
 export { default as artifactImport } from './artifactImport'
 export { default as artifactIdentifier } from './artifactIdentifier'
+export { default as preprocessorTest } from './preprocessorTest'
 export type { EmbeddedGraphqlDocument } from './walkTaggedDocuments'
